pyproject.toml
[project]
name = "dev-uwp-scannerapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.117.1",
    "google-genai>=1.38.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.36.1",
]
